requirements.txt
confluent-kafka>=2.3.0
pyspark==3.5.5
delta-spark==3.2.0
